+++ src/main/java/units/shooter_developers/Projectile.java
@@ -6,8 +6,8 @@ import javafx.util.Pair;
 
 public class Projectile extends DynamicObject {
 
-    private int _biasX;
-    private int _biasY;
+    private double _biasX;
+    private double _biasY;
     private final String Owner;
 
     /* Constructor */
@@ -64,10 +64,10 @@ public class Projectile extends DynamicObject {
     private void setInitialAndTranslateDirection(Direction D) {
         _biasX = _biasY =0;
         switch (D) {
-            case UP    ->  { set_biases(+(int)(getActualWidth()),-(int) (getActualHeight()/2)); set_deltaY(get_deltaY()- get_speed());}
-            case DOWN  ->  { set_biases(+(int)(getActualWidth()),+ (int)(getActualHeight()*2)); set_deltaY(get_deltaY()+ get_speed());}
-            case LEFT  ->  { set_biases(( 0), +(int) getActualHeight()/2);set_deltaX(get_deltaX()-get_speed());}
-            case RIGHT ->  { set_biases((+(int)(getActualWidth()*2)), getActualHeight()/2);set_deltaX(get_deltaX()+get_speed());}
+            case UP    ->  { set_biases(+(getActualWidth()),-(getActualHeight()/2)); set_deltaY(get_deltaY()- get_speed());}
+            case DOWN  ->  { set_biases(+(getActualWidth()),+ (getActualHeight()*2)); set_deltaY(get_deltaY()+ get_speed());}
+            case LEFT  ->  { set_biases(( 0), + getActualHeight()/2);set_deltaX(get_deltaX()-get_speed());}
+            case RIGHT ->  { set_biases((+(getActualWidth()*2)), getActualHeight()/2);set_deltaX(get_deltaX()+get_speed());}
         }
     }
     private double get_biased_y_position(Sprite S) { return S.getFutureY() + _biasY; }
@@ -78,17 +78,17 @@ public class Projectile extends DynamicObject {
 
     private Coordinates getBiasedStartingPosition(Sprite S) { return new Coordinates(get_biased_x_position(S), get_biased_y_position(S)); }
 
-    private void set_biases(int biasX, int biasY)
+    private void set_biases(double biasX, double biasY)
     {
         set_biasX(biasX);
         set_biasY(biasY);
     }
 
-    public void set_biasX(int biasX) {
+    public void set_biasX(double biasX) {
         this._biasX = biasX;
     }
 
-    public void set_biasY(int biasY) {
+    public void set_biasY(double biasY) {
         this._biasY = biasY;
     }
